Created an "only:: html" directive to include a link to invest documentation on the HTML build side.
Fixes issue 4

Creates a link at the *top* of the HTML page that has a link to the pdf.  Unfortuantely, Sphinx's only::html directive is a little buggy and isn't correctly interpreting headers to stick this in the TOC.  I've moved the PDF link to the top so it's obvious to a user who looks at the first page, but they won't see anything in the TOC.

I hope this is an acceptable alternative to having two index files (one for html one for PDF).