Update issue 2363
Making subregions a required input the sneaky way- by just removing the checkbox. The if/else logic still exists within the core, I'm just not making it uptional anymore, so the args input will always exist.

Additionally, changing the UI for hra_core to achieve this. It's no longer a hideableFileEntry, now it's just a file.

Also, fixing the user's guide so that it no longer says it's an optional input.