Add verbiage about default costs in US$
From a Slack conversation with Rob Griffin: "The transmission cost is specified in US dollars (table 3 and the formula above it). This function is hard coded into the model, though the function's parameters are modifiable through the "global wind energy parameters" input. So in the user's guide we'd need to flag that the default parameters for transmission cost are in US$ (millions now, but that could be changed in the ticket above), and they could be changed to whatever currency the user likes by modifying the global wind energy parameters table using an appropriate conversion rate."