Update the citation and contributors list
Also removing the acknowledgments section and adding a css file so that
the contributors names could be in list format without including the
bullets. I thought about adding the css into the `adctheme.css` but
wasn't sure if we were going to move away from this, it might be easier
to just maintain a custom css.