Rewriting confusing conditional under step 4.
The way this conditional has been expressed is really ambiguous and has
created a lot of work for me over email and slack to try to clarify with
users and our science staff.  So I'm removing it and moving the helpful
bits to other subsections to hopefully clarify the intent of a given
reclassification method and clarify what the model is doing.