Handling multiple tags on a single commit.
Git (and thus setuptools_scm) has a weird case where `git describe`
lists the first tag that points to a given commit if there are multiple
tags pointing to that commit.  I do not know if this is guaranteed.

If there are multiple tags on the commit, we now take the latest
available tag.  This should work great for the live builds.  The only
case where this doesn't work out so well is if/when we have checked out
an earlier build and then the UG version will say that it's actually
from a later build.